Move constant to util.py
